1-Product/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
attrs==24.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
dataframe_image==0.2.6
defusedxml==0.7.1
fastjsonschema==2.21.1
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
html2image==2.0.5
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mistune==3.0.2
more-itertools==10.5.0
multidict==6.1.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
playwright==1.49.1
propcache==0.2.1
pyee==12.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
soupsieve==2.6
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
yarl==1.18.3
